+++ rollup.config.js
@@ -7,7 +7,7 @@ import license from 'rollup-plugin-license';
 const ALGORITHMS = [
   'adler32', 'argon2', 'bcrypt', 'blake2b', 'blake2s', 'blake3', 'crc32', 'hmac', 'keccak', 'md4', 'md5',
   'pbkdf2', 'ripemd160', 'scrypt', 'sha1', 'sha3', 'sha224', 'sha256', 'sha384', 'sha512',
-  'sm3', 'whirlpool', 'xxhash3', 'xxhash32', 'xxhash64',
+  'sm3', 'whirlpool', 'xxhash32', 'xxhash64', 'xxhash3', 'xxhash128',
 ];
 
 const TERSER_CONFIG = {